driftlock_sim/requirements.txt
numpy>=1.20
scipy>=1.7
matplotlib>=3.4
pyyaml>=5.4
imageio>=2.34
